Remove AbstractMHQNagDialog_NEW and update references
Replaced AbstractMHQNagDialog_NEW with AbstractMHQNagDialog to clean up redundant code and simplify the class hierarchy. Updated all dependent nag dialog classes to reference AbstractMHQNagDialog for consistency and maintainability.
